test: replace compact tests with overflow redirect integration tests
